+++ transform.jl
@@ -5,7 +5,7 @@ include("circus.jl")
 
 using Images
 
-function prepare_transform(input::Image{Float64}, angle_stepsize::Uint,
+function prepare_transform(input::Image{Float32}, angle_stepsize::Uint,
                            orthonormal::Bool)
     # Orthonormal P-functionals need a stretched image in order to ensure a
     # square sinogram
@@ -24,13 +24,13 @@ function prepare_transform(input::Image{Float64}, angle_stepsize::Uint,
     return input
 end
 
-function get_transform(input::Image{Float64},
+function get_transform(input::Image{Float32},
     tfunctionals::Vector{TFunctionalWrapper},
     pfunctionals::Vector{PFunctionalWrapper},
     angle_stepsize::Uint, orthonormal::Bool, write_data::Bool)
     # Process all T-functionals
     print_debug("Calculating sinograms for given T-functionals\n")
-    sinograms::Vector{Image{Float64}} = getSinograms(input, angle_stepsize,
+    sinograms::Vector{Image{Float32}} = getSinograms(input, angle_stepsize,
                                            tfunctionals)
     for t in 1:length(tfunctionals)
         tfunctional = tfunctionals[t]
